Feat: 회원 챌린지 단일 조회 API 요청시 참여중인 인원 정보 응답 추가

+++ src/main/java/com/example/savingsalt/challenge/repository/MemberChallengeRepository.java
@@ -17,7 +17,7 @@ public interface MemberChallengeRepository extends JpaRepository<MemberChallenge
     List<MemberChallengeEntity> findAllByChallengeEntity(ChallengeEntity challengeEntity);
 
     @Query("SELECT COUNT(c) FROM MemberChallengeEntity c WHERE c.challengeStatus = :status AND c.challengeEntity = :entity")
-    long countByChallengeStatusAndChallengeEntity(
+    Integer countByChallengeStatusAndChallengeEntity(
         @Param("status") MemberChallengeEntity.ChallengeStatus challengeStatus,
         @Param("entity") ChallengeEntity challengeEntity);
